Extract mainClue and put on side in wide screens
